# Node-Server/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification - Sistema de Controle de Estoque
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Fornecedor {
  id: number;
  nomeEmpresa: string;
  cnpj: string;
  endereco: string;
  telefone: string;
  email: string;
  contatoPrincipal: string;
  createdAt: string;
}

export interface CreateFornecedorBody {
  nomeEmpresa: string;
  cnpj: string;
  endereco: string;
  telefone: string;
  email: string;
  contatoPrincipal: string;
}

export interface Produto {
  id: number;
  nome: string;
  /** @nullable */
  codigoBarras?: string | null;
  descricao: string;
  quantidadeEstoque: number;
  categoria: string;
  /** @nullable */
  dataValidade?: string | null;
  /** @nullable */
  imagemUrl?: string | null;
  createdAt: string;
}

export interface ProdutoComFornecedores {
  id: number;
  nome: string;
  /** @nullable */
  codigoBarras?: string | null;
  descricao: string;
  quantidadeEstoque: number;
  categoria: string;
  /** @nullable */
  dataValidade?: string | null;
  /** @nullable */
  imagemUrl?: string | null;
  createdAt: string;
  fornecedores: Fornecedor[];
}

export interface CreateProdutoBody {
  nome: string;
  /** @nullable */
  codigoBarras?: string | null;
  descricao: string;
  quantidadeEstoque?: number;
  categoria: string;
  /** @nullable */
  dataValidade?: string | null;
  /** @nullable */
  imagemUrl?: string | null;
}

export interface AssociarFornecedorBody {
  fornecedorId: number;
}

export interface EstoqueStats {
  totalProdutos: number;
  totalFornecedores: number;
  produtosEstoqueBaixo: number;
  categorias: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListProdutosParams = {
  categoria?: string;
};