# Node-Server/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification - Sistema de Controle de Estoque
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Listar fornecedores
 */
export const ListFornecedoresResponseItem = zod.object({
  id: zod.number(),
  nomeEmpresa: zod.string(),
  cnpj: zod.string(),
  endereco: zod.string(),
  telefone: zod.string(),
  email: zod.string(),
  contatoPrincipal: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListFornecedoresResponse = zod.array(ListFornecedoresResponseItem);

/**
 * @summary Cadastrar fornecedor
 */
export const CreateFornecedorBody = zod.object({
  nomeEmpresa: zod.string(),
  cnpj: zod.string(),
  endereco: zod.string(),
  telefone: zod.string(),
  email: zod.string(),
  contatoPrincipal: zod.string(),
});

/**
 * @summary Buscar fornecedor
 */
export const GetFornecedorParams = zod.object({
  id: zod.coerce.number(),
});

export const GetFornecedorResponse = zod.object({
  id: zod.number(),
  nomeEmpresa: zod.string(),
  cnpj: zod.string(),
  endereco: zod.string(),
  telefone: zod.string(),
  email: zod.string(),
  contatoPrincipal: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Atualizar fornecedor
 */
export const UpdateFornecedorParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateFornecedorBody = zod.object({
  nomeEmpresa: zod.string(),
  cnpj: zod.string(),
  endereco: zod.string(),
  telefone: zod.string(),
  email: zod.string(),
  contatoPrincipal: zod.string(),
});

export const UpdateFornecedorResponse = zod.object({
  id: zod.number(),
  nomeEmpresa: zod.string(),
  cnpj: zod.string(),
  endereco: zod.string(),
  telefone: zod.string(),
  email: zod.string(),
  contatoPrincipal: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Excluir fornecedor
 */
export const DeleteFornecedorParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Listar produtos
 */
export const ListProdutosQueryParams = zod.object({
  categoria: zod.coerce.string().optional(),
});

export const ListProdutosResponseItem = zod.object({
  id: zod.number(),
  nome: zod.string(),
  codigoBarras: zod.string().nullish(),
  descricao: zod.string(),
  quantidadeEstoque: zod.number(),
  categoria: zod.string(),
  dataValidade: zod.string().nullish(),
  imagemUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListProdutosResponse = zod.array(ListProdutosResponseItem);

/**
 * @summary Cadastrar produto
 */
export const CreateProdutoBody = zod.object({
  nome: zod.string(),
  codigoBarras: zod.string().nullish(),
  descricao: zod.string(),
  quantidadeEstoque: zod.number().optional(),
  categoria: zod.string(),
  dataValidade: zod.string().nullish(),
  imagemUrl: zod.string().nullish(),
});

/**
 * @summary Estatísticas de estoque
 */
export const GetProdutoStatsResponse = zod.object({
  totalProdutos: zod.number(),
  totalFornecedores: zod.number(),
  produtosEstoqueBaixo: zod.number(),
  categorias: zod.number(),
});

/**
 * @summary Buscar produto
 */
export const GetProdutoParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProdutoResponse = zod.object({
  id: zod.number(),
  nome: zod.string(),
  codigoBarras: zod.string().nullish(),
  descricao: zod.string(),
  quantidadeEstoque: zod.number(),
  categoria: zod.string(),
  dataValidade: zod.string().nullish(),
  imagemUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  fornecedores: zod.array(
    zod.object({
      id: zod.number(),
      nomeEmpresa: zod.string(),
      cnpj: zod.string(),
      endereco: zod.string(),
      telefone: zod.string(),
      email: zod.string(),
      contatoPrincipal: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Atualizar produto
 */
export const UpdateProdutoParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProdutoBody = zod.object({
  nome: zod.string(),
  codigoBarras: zod.string().nullish(),
  descricao: zod.string(),
  quantidadeEstoque: zod.number().optional(),
  categoria: zod.string(),
  dataValidade: zod.string().nullish(),
  imagemUrl: zod.string().nullish(),
});

export const UpdateProdutoResponse = zod.object({
  id: zod.number(),
  nome: zod.string(),
  codigoBarras: zod.string().nullish(),
  descricao: zod.string(),
  quantidadeEstoque: zod.number(),
  categoria: zod.string(),
  dataValidade: zod.string().nullish(),
  imagemUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Excluir produto
 */
export const DeleteProdutoParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Associar fornecedor ao produto
 */
export const AssociarFornecedorParams = zod.object({
  id: zod.coerce.number(),
});

export const AssociarFornecedorBody = zod.object({
  fornecedorId: zod.number(),
});

/**
 * @summary Desassociar fornecedor do produto
 */
export const DesassociarFornecedorParams = zod.object({
  id: zod.coerce.number(),
  fornecedorId: zod.coerce.number(),
});

export const DesassociarFornecedorResponse = zod.object({
  id: zod.number(),
  nome: zod.string(),
  codigoBarras: zod.string().nullish(),
  descricao: zod.string(),
  quantidadeEstoque: zod.number(),
  categoria: zod.string(),
  dataValidade: zod.string().nullish(),
  imagemUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  fornecedores: zod.array(
    zod.object({
      id: zod.number(),
      nomeEmpresa: zod.string(),
      cnpj: zod.string(),
      endereco: zod.string(),
      telefone: zod.string(),
      email: zod.string(),
      contatoPrincipal: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});