# Node-Server/lib/api-zod/src/generated/types/produtoComFornecedores.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification - Sistema de Controle de Estoque
 * OpenAPI spec version: 0.1.0
 */
import type { Fornecedor } from "./fornecedor";

export interface ProdutoComFornecedores {
  id: number;
  nome: string;
  /** @nullable */
  codigoBarras?: string | null;
  descricao: string;
  quantidadeEstoque: number;
  categoria: string;
  /** @nullable */
  dataValidade?: string | null;
  /** @nullable */
  imagemUrl?: string | null;
  createdAt: Date;
  fornecedores: Fornecedor[];
}